format_template_with_optional_keys moved to lib
